+++ Estimation/Merge/MergeStrategyAbstract.h
@@ -15,14 +15,14 @@ namespace Merge
 		typedef std::vector<size_t> ul_list_t;
 
 	private:
-		const int _min_genes_before_merge;
-		const int _min_genes_after_merge;
+		const size_t _min_genes_before_merge;
+		const size_t _min_genes_after_merge;
 
 	protected:
 		virtual ul_list_t merge_inited(Estimation::CellsDataContainer &container) const = 0;
 
 	public:
-		MergeStrategyAbstract(unsigned min_genes_before_merge, unsigned min_genes_after_merge);
+		MergeStrategyAbstract(size_t min_genes_before_merge, size_t min_genes_after_merge);
 
 		ul_list_t merge(Estimation::CellsDataContainer &container);
 
@@ -30,8 +30,8 @@ namespace Merge
 		virtual void release();
 		virtual std::string merge_type() const = 0;
 
-		int min_genes_before_merge() const;
-		int min_genes_after_merge() const;
+		size_t min_genes_before_merge() const;
+		size_t min_genes_after_merge() const;
 	};
 }
 }
