+++ Estimation/Merge/PoissonRealBarcodesMergeStrategy.cpp
@@ -9,9 +9,9 @@ namespace Merge
 {
 	PoissonRealBarcodesMergeStrategy::PoissonRealBarcodesMergeStrategy(const PoissonTargetEstimator &target_estimator,
 	                                                                   const barcodes_parser_ptr barcodes_parser,
-																	   unsigned min_genes_before_merge,
-																	   unsigned min_genes_after_merge,
-																	   unsigned max_merge_edit_distance)
+	                                                                   size_t min_genes_before_merge,
+	                                                                   size_t min_genes_after_merge,
+	                                                                   unsigned max_merge_edit_distance)
 		: RealBarcodesMergeStrategy(barcodes_parser, min_genes_before_merge, min_genes_after_merge,
 		                            max_merge_edit_distance, 0)
 		, _target_estimator(target_estimator)
