+++ Estimation/Merge/PoissonRealBarcodesMergeStrategy.h
@@ -31,8 +31,8 @@ namespace Merge
 
 	public:
 		PoissonRealBarcodesMergeStrategy(const PoissonTargetEstimator &target_estimator,
-		                                 const barcodes_parser_ptr barcodes_parser, unsigned min_genes_before_merge,
-		                                 unsigned min_genes_after_merge, unsigned max_merge_edit_distance);
+		                                 const barcodes_parser_ptr barcodes_parser, size_t min_genes_before_merge,
+		                                 size_t min_genes_after_merge, unsigned max_merge_edit_distance);
 
 		virtual std::string merge_type() const override;
 	};
