+++ Estimation/Merge/RealBarcodesMergeStrategy.h
@@ -51,7 +51,7 @@ namespace Merge
 
 	public:
 		RealBarcodesMergeStrategy(barcodes_parser_ptr barcodes_parser,
-		                          unsigned min_genes_before_merge, unsigned min_genes_after_merge,
+		                          size_t min_genes_before_merge, size_t min_genes_after_merge,
 		                          unsigned max_merge_edit_distance, double min_merge_fraction);
 
 		virtual std::string merge_type() const override;
